Fixed the bug, that one account with the same address was created twice
Fixed the bug by first inserting all new contracts, that are created in the block, and then after this, creating the remaining external accounts: 
First we need to create all contracts, since we have to wait for the transaction receipt. Otherwise we potentially violate the uniqueness constraint. That could happen when in one transaction the account is created and in the next transaction of the block it is called immediately.